Fix: Validation working well \n Feature: Moved to routes based on role and reduces endpoint according to the needs \n New: Added the Postman Collection

+++ middlewares/index.middleware.js
@@ -2,10 +2,17 @@ const User = require("../models/user.model")
 const Course = require("../models/course.model")
 const Article = require("../models/article.model")
 const Question = require("../models/question.model")
-const bcrypt = require("bcrypt");
-const { application } = require("express");
+const Delete = require("../delete/deleted.user.model")
 
 // user
+module.exports.user = (callback) => {
+    User.find({}, callback)
+}
+
+module.exports.findUser = (userEmail, callback) => {
+    User.find({email: userEmail.email}).populate("courses").exec(callback)
+}
+
 module.exports.newUser = async (user, callback) => {
     User.create(user, (err, newUser) => {
         if (err) {
@@ -34,7 +41,9 @@ module.exports.userUpdate = (user, callback) => {
             "country": user.country,
             "state": user.state,
             "DP": user.DP,
-            "password": user.password
+            "password": user.password,
+            "role": user.role,
+            "courses" : user.courses
         },
         {new: true, upsert: true},
         callback
@@ -42,23 +51,35 @@ module.exports.userUpdate = (user, callback) => {
 }
 
 module.exports.delUser = (user, callback) => {
-    User.findByIdAndDelete(
+    User.findById(
         user.id,
-        callback
+       async (err, foundUser) => {
+            if (err) {
+                return err
+            } else {
+                 const delUser = {
+                    deleteUser: foundUser.toJSON()
+                }
+               await Delete.create(delUser, (err, newDelUser) => {
+                    if (err) {
+                        return err
+                    } else {
+                        newDelUser.save()
+                    }
+                })
+               await User.findByIdAndDelete(user.id, callback)
+            }
+        }
     )
 }
 
 // course
-module.exports.delCourse = (course, callback) => {
-    Course.findByIdAndDelete(
-        course.id,
-        callback
-    )
+module.exports.course = (callback) => {
+    Course.find({}, callback)
 }
 
-module.exports.courses = (course, callback) => {
-    const newCourse = new Course(course)
-    newCourse.save(callback)
+module.exports.findCourse = (courseTitle, callback) => {
+    Course.find({title: courseTitle.title}, callback)
 }
 
 module.exports.courseUpdate = (course, callback) => {
@@ -71,30 +92,82 @@ module.exports.courseUpdate = (course, callback) => {
             "durationPerQuestion": course.durationPerQuestion,
             "totalQuestion": course.totalQuestion,
             "price": course.price,
-            "headline": course.headline
+            "headline": course.headline,
+            "isPaid": course.isPaid,
+            "tutor": course.tutor,
+            "isPrivate": course.isPrivate,
+            "numTeachers": course.numTeachers,
+            "priceCurrency": course.priceCurrency,
+            "primaryCategory": course.primaryCategory,
+            "subCategory": course.subCategory,
+            "language": course.language,
+            "statusLabel": course.statusLabel
         },
         {new: true, upsert: true},
         callback
     )
 }
 
-// article
-module.exports.delArticle = (article, callback) => {
-    Article.findByIdAndDelete(
-        article.idA,
+module.exports.delCourse = async (course, callback) => {
+    await Article.find(
+        {courseID: course.id}, async (err, foundArticles) => {
+            if (err) {
+                return err
+            } else {
+                for (const _id in foundArticles) {
+                    if (foundArticles.hasOwnProperty(_id)) {
+                        const element = foundArticles[_id];
+                       await Article.findByIdAndDelete(
+                            element._id,
+                            (err, data) => {
+                                if (err) {
+                                    return err
+                                } else {
+                                    console.log(data)
+                                }
+                            }
+                        )
+                    }
+                }
+            }
+        }
+    )
+
+    await Question.find(
+        {courseID: course.id}, async (err, foundQuestions) => {
+            if (err) {
+                return err
+            } else {
+                for (const _id in foundQuestions) {
+                    if (foundQuestions.hasOwnProperty(_id)) {
+                        const element = foundQuestions[_id];
+                        await Question.findByIdAndDelete(
+                            element._id,
+                            (err, data) => {
+                                if (err) {
+                                    return err
+                                } else {
+                                    console.log(data)
+                                }
+                            }
+                        )
+                    }
+                }
+            }
+        }
+    ),
+
+    await Course.findByIdAndDelete(
+        course.id,
         callback
     )
 }
 
-module.exports.newArticle = (article, callback) => {
-    Article.create(article, (err, newArticle) => {
-        if (err) {
-            console.log(`${err} from create article`)
-            console.log(err)
-        } else {
-            newArticle.save(callback)
-        }
-    })
+// article
+module.exports.getCourseArticle = (article, callback) => {
+    Article.find({courseID: article.courseID},
+        callback
+    )
 }
 
 module.exports.articleUpdate = (article, callback) => {
@@ -110,8 +183,9 @@ module.exports.articleUpdate = (article, callback) => {
     )
 }
 
-module.exports.getCourseArticle = (article, callback) => {
-    Article.find({courseID: article.courseID},
+module.exports.delArticle = (article, callback) => {
+    Article.findByIdAndDelete(
+        article.idA,
         callback
     )
 }
@@ -123,17 +197,6 @@ module.exports.getCourseQuestion = (question, callback) => {
     )
 }
 
-module.exports.newQuestion = (question, callback) => {
-    Question.create(question, (err, newQuestion) => {
-        if (err) {
-            console.log(`${err} from create question`)
-            return err
-        } else {
-            newQuestion.save(callback)
-        }
-    })
-}
-
 module.exports.questionUpdate = (question, callback) => {
     Question.findByIdAndUpdate(
         question.idQ,
@@ -143,7 +206,8 @@ module.exports.questionUpdate = (question, callback) => {
             "option2": question.option2,
             "option3": question.option3,
             "option4": question.option4,
-            "correctAnswer": question.correctAnswer
+            "correctAnswer": question.correctAnswer,
+            "tutor": question.tutor
         },
         {new: true, upsert: true},
         callback
@@ -154,32 +218,5 @@ module.exports.delQuestion = (question, callback) => {
     Question.findByIdAndDelete(
         question.idQ,
         callback
-    )
-}
-
-// application
-module.exports.newApplication = async (application, callback) => {
-    User.findById(application._id, (err, foundUser) => {
-        if (err) {
-            console.log(`${err} from application find user`)
-            return err
-        } else {
-            foundUser.courses.push(application.appID)
-            foundUser.save(callback)
-        }
-    })
-}
-
-module.exports.delApplication = async (application, callback) => {
-    await User.findById(
-        application.id, (err, foundUser) => {
-            if (err) {
-                return err
-            } else {
-                foundUser.updateOne({"courses": foundUser.courses.filter((course) => { 
-                    return course != application.appID
-                })}, callback)
-            }
-        }
     )
 }
